Eliminada NN Distribution_convocatory
Se ha borrado la relación existente entre convocatoria, Alumno y
Curso académico. Se ha sustituido por dos relaciones 1N.
