--- src/AppBundle/Repository/SchoolYear_convocatoryRepository.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace AppBundle\Repository;
-
-/**
- * SchoolYear_convocatoryRepository
- *
- * This class was generated by the Doctrine ORM. Add your own custom
- * repository methods below.
- */
-class SchoolYear_convocatoryRepository extends \Doctrine\ORM\EntityRepository
-{
-    public function getSchoolYearConvocatories()
-    {
-        return $this->findAll();
-    }
-}
